detectText/requirements.txt
loguru==0.6.0
MouseInfo==0.1.3
notify-py==0.3.42
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.2.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
python-dotenv==1.0.1
pytweening==1.2.0
rubicon-objc==0.4.7
